Refactored target builder to move some logic outside of the builder, into smaller fns
