+++ packages/target-graph/src/TargetGraphBuilder.ts
@@ -1,5 +1,7 @@
 import { createDependencyMap } from "workspace-tools/lib/graph/createDependencyMap";
 import { getPackageAndTask, getStartTargetId, getTargetId } from "./targetId";
+import { prioritize } from "./prioritize";
+import { expandDepSpecs } from "./expandDepSpecs";
 
 import path from "path";
 
@@ -22,9 +24,6 @@ import type { TargetConfig } from "./types/TargetConfig";
  * ```
  */
 export class TargetGraphBuilder {
-  /** Cached transitive task dependency */
-  private cachedTransitiveTaskDependencies: Map<string, "walk-in-progress" | Set<string>> = new Map();
-
   /** A map of targets - used internally for looking up generated targets from the target configurations */
   private targets: Map<string, Target> = new Map();
 
@@ -103,183 +102,6 @@ export class TargetGraphBuilder {
   }
 
   /**
-   * Generates new `Target`, indexed by the id based on a new target configuration.
-   *
-   * @param id
-   * @param targetDefinition
-   */
-  addTargetConfig(id: string, config: TargetConfig = {}): void {
-    // Generates a target definition from the target config
-    if (id.startsWith("//") || id.startsWith("#")) {
-      const target = this.createGlobalTarget(id, config);
-      this.targets.set(target.id, target);
-    } else if (id.includes("#")) {
-      const { packageName, task } = getPackageAndTask(id);
-      const target = this.createPackageTarget(packageName!, task, config);
-      this.targets.set(target.id, target);
-    } else {
-      const packages = Object.keys(this.packageInfos);
-      for (const packageName of packages) {
-        const task = id;
-        const target = this.createPackageTarget(packageName!, task, config);
-        this.targets.set(target.id, target);
-      }
-    }
-  }
-
-  /**
-   * Expands the dependency graph by adding all transitive dependencies of the given targets.
-   */
-  private expandDepSpecs() {
-    /**
-     * Adds a dependency in the form of [from, to] to the dependency list.
-     * @param from
-     * @param to
-     */
-    const addDependency = (from: string, to: string) => {
-      this.dependencies.push([from, to]);
-
-      this.targets.get(from)!.dependents.push(to);
-      this.targets.get(to)?.dependencies.push(from);
-    };
-
-    /**
-     * Finds all transitive dependencies, given a task and optionally a dependency list.
-     * @param task
-     * @param dependencies
-     * @returns
-     */
-    const findDependenciesByTask = (task: string, dependencies?: string[]) => {
-      if (!dependencies) {
-        return targets.filter((needle) => needle.task === task).map((needle) => needle.id);
-      }
-
-      return targets
-        .filter((needle) => {
-          const { task: needleTask, packageName: needlePackageName } = needle;
-          return needleTask === task && dependencies.some((depPkg) => depPkg === needlePackageName);
-        })
-        .map((needle) => needle.id);
-    };
-
-    const targets = [...this.targets.values()];
-
-    for (const target of targets) {
-      const { depSpecs, packageName, id: to } = target;
-
-      // Always start with a root node with a special "START_TARGET_ID"
-      // because any node could potentially be part of the entry point in building the scoped target subgraph
-      this.dependencies.push([getStartTargetId(), to]);
-
-      // Skip any targets that have no "deps" specified
-      if (!depSpecs || depSpecs.length === 0) {
-        continue;
-      }
-
-      /**
-       * Now for every deps defined, we need to "interpret" it based on the syntax:
-       * - for any deps like package#task, we simply add the singular dependency (source could be a single package or all packages)
-       * - for anything that starts with a "^", we add the package-tasks according to the topological package graph
-       *    NOTE: in a non-strict mode (TODO), the dependencies can come from transitive task dependencies
-       * - for anything that starts with a "^^", we add the package-tasks from the transitive dependencies in the topological
-       *    package graph.
-       * - for {"pkgA#task": ["dep"]}, we interpret to add "pkgA#dep"
-       * - for anything that is a string without a "^", we treat that string as the name of a task, adding all targets that way
-       *    NOTE: in a non-strict mode (TODO), the dependencies can come from transitive task dependencies
-       *
-       * We interpret anything outside of these conditions as invalid
-       */
-      for (const dependencyTargetId of depSpecs) {
-        if (dependencyTargetId.includes("#")) {
-          // id's with a # are package-task dependencies, or global
-          // therefore, we must use getPackageAndTask() & getTargetId() to normalize the target id
-          // (e.g. "build": ["build-tool#build"])
-          const { packageName, task } = getPackageAndTask(dependencyTargetId);
-          const normalizedDependencyTargetId = getTargetId(packageName, task);
-          addDependency(normalizedDependencyTargetId, to);
-        } else if (dependencyTargetId.startsWith("^^") && packageName) {
-          // Transitive depdency (e.g. bundle: ['^^transpile'])
-          const depTask = dependencyTargetId.substring(2);
-          const targetDependencies = [...(this.getTransitiveGraphDependencies(packageName) ?? [])];
-          const dependencyTargetIds = findDependenciesByTask(depTask, targetDependencies);
-          for (const from of dependencyTargetIds) {
-            addDependency(from, to);
-          }
-        } else if (dependencyTargetId.startsWith("^") && packageName) {
-          // Topological dependency (e.g. build: ['^build'])
-          const depTask = dependencyTargetId.substring(1);
-          const targetDependencies = [...(this.dependencyMap.dependencies.get(packageName) ?? [])];
-          const dependencyTargetIds = findDependenciesByTask(depTask, targetDependencies);
-          for (const from of dependencyTargetIds) {
-            addDependency(from, to);
-          }
-        } else if (packageName) {
-          // Add dependency on a specific package and given task name as dependency
-          // (e.g. bundle: ['build'])
-          const task = dependencyTargetId;
-          if (this.targets.has(getTargetId(packageName, task))) {
-            addDependency(getTargetId(packageName, task), to);
-          }
-        } else if (!dependencyTargetId.startsWith("^")) {
-          // Global dependency - add all targets that match task name as dependency
-          // (e.g. "#bundle": ['build'])
-          const task = dependencyTargetId;
-          const dependencyIds = findDependenciesByTask(task);
-          for (const dependencyId of dependencyIds) {
-            addDependency(dependencyId, to);
-          }
-        } else {
-          throw new Error(`invalid pipeline config detected: ${target.id}, packageName: ${packageName}, dep: ${dependencyTargetId}`);
-        }
-      }
-    }
-  }
-
-  /**
-   * Gets a list of package names that are direct or indirect dependencies of rootPackageName in this.graph,
-   * and caches them on the Pipeline.
-   *
-   * For example, this is useful for a bundling target that depends on all transitive dependencies to have been built.
-   *
-   * @param packageName the root package to begin walking from
-   */
-  private getTransitiveGraphDependencies(packageName: string): Set<string> {
-    const cachedResult = this.cachedTransitiveTaskDependencies.get(packageName);
-    if (cachedResult) {
-      return cachedResult === "walk-in-progress"
-        ? // There is a recursive walk over this set of dependencies in progress.
-          // If we hit this case, that means that a dependency of this package depends on it.
-          //
-          // In this case we return an empty set to omit this package and it's downstream from its
-          // dependency
-          new Set()
-        : // we already computed this for this package, return the cached result.
-          cachedResult;
-    } else {
-      // No cached result. Compute now with a recursive walk
-
-      // mark that we are traversing this package to prevent infinite recursion
-      // in cases of circular dependencies
-      this.cachedTransitiveTaskDependencies.set(packageName, "walk-in-progress");
-
-      const immediateDependencies = [...(this.dependencyMap.dependencies.get(packageName) ?? [])];
-
-      // build the set of transitive dependencies by recursively walking the
-      // immediate dependencies' dependencies.
-      const transitiveDepSet = new Set<string>(immediateDependencies);
-      for (const immediateDependency of immediateDependencies) {
-        for (const transitiveSubDependency of this.getTransitiveGraphDependencies(immediateDependency)) {
-          transitiveDepSet.add(transitiveSubDependency);
-        }
-      }
-
-      // Cache the result and return
-      this.cachedTransitiveTaskDependencies.set(packageName, transitiveDepSet);
-      return transitiveDepSet;
-    }
-  }
-
-  /**
    * Filters out targets that are not part of the entry point.
    * @param tasks
    * @param scope
@@ -354,37 +176,30 @@ export class TargetGraphBuilder {
   }
 
   /**
-   * Priorities for a target is actually the MAX of all the priorities of the targets that depend on it.
+   * Generates new `Target`, indexed by the id based on a new target configuration.
+   *
+   * @param id
+   * @param targetDefinition
    */
-  setupPriorities() {
-    const stack: Target[] = [];
-
-    // start with the leaves - the children of the all targets
-    for (const target of this.targets.values()) {
-      if (target.dependencies.length === 0) {
-        stack.push(target);
-      }
-    }
-
-    // fill every target with the max priority of its dependents
-    while (stack.length > 0) {
-      const currentTarget = stack.pop()!;
-      const currentPriority = currentTarget.priority;
-
-      for (const dependentId of currentTarget.dependents) {
-        const dependent = this.targets.get(dependentId)!;
-        dependent.priority = Math.max(dependent.priority ?? 0, currentPriority ?? 0);
+  addTargetConfig(id: string, config: TargetConfig = {}): void {
+    // Generates a target definition from the target config
+    if (id.startsWith("//") || id.startsWith("#")) {
+      const target = this.createGlobalTarget(id, config);
+      this.targets.set(target.id, target);
+    } else if (id.includes("#")) {
+      const { packageName, task } = getPackageAndTask(id);
+      const target = this.createPackageTarget(packageName!, task, config);
+      this.targets.set(target.id, target);
+    } else {
+      const packages = Object.keys(this.packageInfos);
+      for (const packageName of packages) {
+        const task = id;
+        const target = this.createPackageTarget(packageName!, task, config);
+        this.targets.set(target.id, target);
       }
     }
   }
 
-  sortTargetIdsByPriority(targets: Target[]) {
-    return targets
-      .sort((a, b) => {
-        return (b.priority ?? 0) - (a.priority ?? 0);
-      })
-      .map((t) => t.id);
-  }
   /**
    * Builds a scoped target graph for given tasks and packages
    *
@@ -399,8 +214,8 @@ export class TargetGraphBuilder {
    * @returns
    */
   buildTargetGraph(tasks: string[], scope?: string[]) {
-    this.expandDepSpecs();
-    this.setupPriorities();
+    this.dependencies = expandDepSpecs(this.targets, this.dependencyMap);
+    prioritize(this.targets);
 
     const startId = getStartTargetId();
     this.targets.set(startId, {
@@ -424,12 +239,9 @@ export class TargetGraphBuilder {
       }
     }
 
-    const targetIdsByPriority = this.sortTargetIdsByPriority([...subGraphTargets.values()]);
-
     return {
       targets: subGraphTargets,
       dependencies: subGraphEdges,
-      targetIdsByPriority,
     };
   }
 }
